New print function and new odds calculation
Created String boardTokens[][] which stores Strings that correspond to
the graphic (i.e. “LUCKY7”, “BAR”, etc.)
Also adjusted SlotMachineImpl.java, virtualSpint(int num) to receive an
integer from 0 - 30.  This allowed me to adjust the odds of each of the
winning combinations.   It allows a winner to be generated every 7
spins or so.
Also started on boolean isWinner( int x, int y, int z) which determines
whether a particular set of three integers (representing a row) is a
winner.